+++ src/main/java/onboarding/problem1/Game.java
@@ -1,4 +1,4 @@
-package onboarding;
+package onboarding.problem1;
 
 import java.util.stream.Stream;
