+++ src/main/java/onboarding/problem1/PagePair.java
@@ -1,4 +1,4 @@
-package onboarding;
+package onboarding.problem1;
 
 public class PagePair {
