+++ src/main/java/onboarding/problem1/Problem1.java
@@ -1,8 +1,8 @@
-package onboarding;
+package onboarding.problem1;
 
 import java.util.List;
 
-class Problem1 {
+public class Problem1 {
 
     public static int solution(List<Integer> pobi, List<Integer> crong) {
